Update/generalize paths for GFSv17 COM reorg
The upcoming update to GFS v17 will substantially change the COM paths
for the GFS package. When called within global workflow, `$COMINgfs` now
contains the full path to the desired directories. This requires changes
to how the look-behind and look-ahead are done. Now instead of constructing
the path itself, the lookbehind and look-ahead now do a substitution of
`${PDY}/${cyc}` with the new desired values for the other cycles.

Additional changes will be needed to the j-jobs to match the updated
`$COMINgfs` pattern for GFS v17 in order to run this as a seperate system
(as in ops). Also, not all GFS tracker jobs are currently available in the
global workflow, so changes for FSU genesis have not been tested.

Additional changes will also be needed for GEFS v13, which is making similar
changes to COM.

Refs: #1
Refs: NOAA-EMC/global-workflow#761